requirements.txt
Django==1.8.4
django-ckeditor==5.0.2
Pillow==2.9.0
psycopg2==2.6.1
wheel==0.24.0
pyquery==1.2.9
django-suit==0.2.14